requirements.txt
asgiref==3.3.0
click==7.1.2
graphql-core==3.1.2
h11==0.11.0
hupper==1.10.2
Pygments==2.7.1
python-dateutil==2.8.1
six==1.15.0
starlette==0.13.8
strawberry-graphql==0.37.0
typing-extensions==3.7.4.3
uvicorn==0.12.2
